src/seminars/third/tdd/User.java
package seminars.third.tdd;

public class User {

    public String name;
    String password;
    boolean isAdmin;
    public boolean isAuthenticate = false;

    public User(String name, String password, boolean isAdmin) {
        this.name = name;
        this.password = password;
        this.isAdmin = isAdmin;

    }

    //3.6.

    public boolean authenticate(String name, String password) {
        if (name.equals(this.name) && password.equals(this.password)) {
            isAuthenticate = true;
            return true;
        } else {
            isAuthenticate = false;
            return false;
        }
    }

    public void logOut() {
        this.isAuthenticate = false;
    }
}